added the bottom sheet for block atsign
